Refactor k-sort-handle
