+++ panel/src/config/components.js
@@ -239,7 +239,7 @@ import Headline from "@/ui/components/Headline.vue";
 import Icon from "@/components/Misc/Icon.vue";
 import Image from "@/components/Misc/Image.vue";
 import Progress from "@/ui/components/Progress.vue";
-import SortHandle from "@/components/Misc/SortHandle.vue";
+import SortHandle from "@/ui/components/SortHandle.vue";
 import Text from "@/ui/components/Text.vue";
 
 Vue.component("k-draggable", Draggable);
