+++ panel/src/ui/components/SortHandle.vue
